@@ -1,6 +1,6 @@
 {
   "name": "@ducat-unit/runestone",
-  "version": "2.0.1",
+  "version": "2.0.2",
   "description": "Fork of the Magic Eden runestone library.",
   "repository": {
     "type": "git",
@@ -18,6 +18,7 @@
     "dist",
     "src"
   ],
+  "type": "module",
   "exports": {
     ".": {
       "import": "./dist/index.js",
@@ -26,12 +27,12 @@
     }
   },
   "scripts": {
-    "script": "tsx",
+    "script": "tsx --tsconfig ./test/tsconfig.json",
     "test": "npm run script test/tape.ts",
     "build": "./scripts/build.sh",
     "check": "npx tsc --project tsconfig.json --noEmit",
-    "lint": "biome check src/lib src/types test index.ts",
-    "lint:fix": "biome check src/lib src/types test index.ts --apply",
+    "lint": "biome check src/",
+    "lint:fix": "biome check src/ --apply",
     "package": "./scripts/package.sh",
     "release": "./scripts/release.sh",
     "start": "node index.js"
@@ -41,11 +42,15 @@
   "license": "ISC",
   "devDependencies": {
     "@biomejs/biome": "2.3.13",
-    "@types/node": "^20.10.7",
-    "@types/tape": "^5.8.1",
-    "faucet": "^0.0.4",
-    "tape": "^5.9.0",
-    "tsx": "^4.21.0",
-    "typescript": "^5.3.3"
+    "@types/node": "25.0.10",
+    "@types/tape": "5.8.1",
+    "faucet": "0.0.4",
+    "tape": "5.9.0",
+    "tslib": "2.8.1",
+    "tsx": "4.21.0",
+    "typescript": "5.9.3"
+  },
+  "dependencies": {
+    "@vbyte/buff": "^1.1.0"
   }
 }
