@@ -1,2 +1,2 @@
-export * from './lib';
-export * from './types';
+export * from './lib/index.js';
+export * from './types/index.js';
