@@ -1,4 +1,4 @@
-export { u8 } from './u8';
-export { u32 } from './u32';
-export { u64 } from './u64';
-export { u128 } from './u128';
+export { u8 } from './u8.js';
+export { u32 } from './u32.js';
+export { u64 } from './u64.js';
+export { u128 } from './u128.js';
