@@ -1,3 +1,3 @@
-export * from './decode';
-export * from './encode';
-export * from './typeguards';
+export * from './decode.js';
+export * from './encode.js';
+export * from './typeguards.js';
