@@ -1,14 +1,16 @@
+import { Buff } from '@vbyte/buff';
+
 export class SeekBuffer {
   public seekIndex: number = 0;
 
-  constructor(private buffer: Buffer) {}
+  constructor(private buffer: Buff) {}
 
   readUInt8(): number | undefined {
     if (this.isFinished()) {
       return undefined;
     }
 
-    return this.buffer.readUInt8(this.seekIndex++);
+    return this.buffer[this.seekIndex++];
   }
 
   isFinished(): boolean {
