@@ -1,5 +1,5 @@
-import { Option } from './monads';
-import { u128, u64 } from './integer';
+import { Option } from './monads.js';
+import { u128, u64 } from './integer/index.js';
 
 export type Terms = {
   amount: Option<u128>;
